refactor(pages): split RealTrackerViewConnection into its own view-connection module
Pure move — view.ts now holds only RealIndividualTrackerViewService; the
WebSocket connection class (+ message-parse helper) lives in
view-connection.ts. No behaviour change.
